refactor: kafka에서 사용하는 토픽명 변경
