+++ charmander-api/charmander-api-video/src/main/java/org/cm/api/VideoConcatConsumer.java
@@ -6,8 +6,9 @@ import org.cm.common.utils.RandomKeyGenerator;
 import org.cm.domain.task.SceneOutputRepository;
 import org.cm.infra.mediaconvert.queue.AllVideoCombineQueue;
 import org.cm.infra.mediaconvert.queue.WavVideoCombineQueue.VideoSource;
-import org.cm.kafka.OverlayCompleteRecord;
+import org.cm.kafka.KafkaTopicNames;
 import org.cm.kafka.UserMetadata;
+import org.cm.kafka.VideoCompositionRecord;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
@@ -20,8 +21,8 @@ public class VideoConcatConsumer {
     private final AllVideoCombineQueue allVideoCombineQueue;
 
     @Transactional
-    @KafkaListener(topics = "video-concat-task", groupId = "tts-task-group")
-    public void listen(OverlayCompleteRecord record) {
+    @KafkaListener(topics = KafkaTopicNames.VIDEO_COMPOSITION, groupId = "tts-task-group")
+    public void listen(VideoCompositionRecord record) {
         sceneOutputRepository.getById(record.sceneId()).updateSceneId(record.sceneVideoId());
 
         var videoSources = getVideoSources(record);
@@ -35,7 +36,7 @@ public class VideoConcatConsumer {
         );
     }
 
-    private List<VideoSource> getVideoSources(OverlayCompleteRecord record) {
+    private List<VideoSource> getVideoSources(VideoCompositionRecord record) {
         return sceneOutputRepository.findAllByTaskId(record.taskId())
                 .stream()
                 .map(sceneOutput -> new VideoSource(sceneOutput.getSceneFileId(), 1280, 720, 0, 0))
